requirements.txt

# Core Data Science
numpy==2.2.4
pandas==2.2.3
scipy==1.15.2
scikit-learn==1.6.1
statsmodels==0.14.4

# Visualization (Reduced to key libraries)
matplotlib==3.10.1  # Base visualization
plotly==6.0.1       # Interactive visualizations
plotnine = 0.14.5

# Machine Learning & AI
torch==2.6.0        # Deep learning
transformers==4.51.3  # Hugging Face transformers
sentence-transformers==4.1.0  # Sentence embeddings
huggingface-hub==0.30.2  # Model hub access
bertopic==0.17.0    # Topic modeling
# Note: Removed redundant ML packages that are dependencies of others

# LangChain/LangGraph (Consolidated)
langchain-core==0.3.56
langgraph==0.4.1
langsmith==0.3.39
# Note: Removed redundant langgraph packages

# Google API (Consolidated)
google-api-python-client==2.167.0  # Core client for Google APIs
google-auth==2.39.0  # Authentication
google-genai==1.10.0  # Generative AI access
googleapis-common-protos==1.70.0
# Note: Removed older/duplicate Google packages

# JSON Processing (Consolidated)
jsonschema==4.23.0  # JSON schema validation
orjson==3.10.18     # Fast JSON parsing
# Note: Removed redundant JSON libraries

# HTTP Clients (Consolidated)
httpx==0.28.1       # Modern HTTP client
# Note: Removed redundant HTTP clients

# Azure (Kept all, likely for different services)
azure-core==1.33.0
azure-cosmos==4.9.0
azure-functions==1.23.0

# Document Processing
python-docx==1.1.2  # DOCX processing
openpyxl==3.1.5     # Excel processing
python-pptx==1.0.2  # PowerPoint processing
pypdfium2==4.30.1   # PDF processing

# NLP & Text Processing
nltk==3.9.1         # Natural language toolkit
regex==2024.11.6    # Enhanced regular expressions
textblob==0.19.0    # Text processing
lxml==5.3.2         # XML/HTML processing
beautifulsoup4==4.13.4  # HTML parsing

# Docling (Kept all, specialized package)
docling==2.30.0
docling-core==2.27.0
docling-ibm-models==3.4.1
docling-parse==4.0.1

# Image Processing
pillow==11.1.0      # Image processing
opencv-python-headless==4.11.0.86  # Computer vision
easyocr==1.7.2      # OCR capabilities
scikit-image==0.25.2  # Image processing

# Jupyter & Development Environment
ipython==9.1.0      # Interactive Python
jupyter_core==5.7.2  # Jupyter core
ipykernel==6.29.5   # IPython kernel

# Utilities
tqdm==4.67.1        # Progress bars
requests==2.32.3    # HTTP requests
python-dotenv==1.1.0  # Environment variables
pydantic==2.11.3    # Data validation

# Web & API Development
streamlit==1.44.1   # Web app framework

# Data Retrieval & APIs
gdeltdoc==1.12.0    # GDELT document API

# Essential Dependencies (kept as they're likely required by other packages)
typing_extensions==4.13.2
PyYAML==6.0.2